[MOD-14883] string_utils: share the libnu read-ahead model out of rs_token
rs_token pads around libnu's decoder because that decoder reads ahead past the
end of a sequence, and the model of how far it can read — NU_MAX_READAHEAD,
nu_seq_len, tail_may_overread — has lived in rs_token as private items.

The fuzzy walk added next has to hand bytes to the same decoder and so needs
the same guards. Move them into string_utils::libnu and have rs_token use them
from there, so the two callers share one copy: a second copy is one a later fix
to the first would leave behind, and being wrong about how far the decoder
reads is a read past the end of a buffer.

No behaviour change — the items are the same, only their home is.
