+++ src/redisearch_rs/c_wrappers/rs_token/Cargo.toml
@@ -21,6 +21,7 @@ ffi.workspace = true
 query_term.workspace = true
 redis-module.workspace = true
 rqe_wildcard.workspace = true
+string_utils.workspace = true
 workspace_hack.workspace = true
 
 [dev-dependencies]
